Backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
python-dotenv>=0.19.0,<0.20.0
sqlalchemy>=1.4.0,<2.0.0
alembic>=1.7.0,<2.0.0
pytest>=6.2.0,<7.0.0
pytest-asyncio>=0.15.0,<0.16.0
httpx>=0.18.0,<0.19.0